Reorg folders, fixup deps, logs
