+++ api/server.js
@@ -3,19 +3,15 @@
 const beforeExit = require('before-exit')
 const config = require('config')
 
-const events = require('./events')
-const indexer = require('./indexer/src')
-const logger = require('./shared/src/logger')
-const restApi = require('./rest-api/src')
+const events = require('../events')
+const logger = require('../shared/src/logger')
+const restApi = require('../rest-api/src')
 
 logger.info('Starting up...')
 logger.debug('Startup configuration: %j', config)
 
-Promise.all([
-  indexer.start(config),
-  restApi.start(config, { socketio: true })
-    .then(server => events.start(config, server))
-])
+restApi.start(config, { socketio: true })
+  .then(server => events.start(config, server))
   .catch(function (err) {
     logger.error('Terminating on error: %s', err.message)
     logger.debug(err.stack)
