+++ functions/get-best-block/create-function.js
@@ -2,16 +2,16 @@
 
 const { pick } = require('lodash')
 
-const logger = require('../../../shared/logger')
+const logger = require('../../shared/src/logger')
 
 function createGetBestBlock (db) {
   // Return the best parsed block
   function getBestBlock (req, res, next) {
-    logger.verbose('-->', req.url)
+    logger.verbose('--> %s', req.url)
 
     db.getBestBlock()
       .then(function (block) {
-        logger.verbose('<--', block)
+        logger.verbose('<-- %j', block)
         res.json(pick(block, ['number', 'hash', 'totalDifficulty']))
         next()
       })
