+++ functions/get-best-block/index.js
