+++ functions/index.js
@@ -1,0 +1,6 @@
+'use strict'
+
+module.exports = {
+  createGetAddrTxs: require('./get-address-transactions').create,
+  createGetBestBlock: require('./get-best-block').create
+}
