+++ indexer/src/pub.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const getPubSub = require('../../shared/pubsub')
+const getPubSub = require('../../shared/src/pubsub')
 
 function createPub (config) {
   // Create a pubsub connection
