+++ package.json
@@ -6,7 +6,7 @@
   "main": "server.js",
   "repository": "autonomoussoftware/ethereum-transaction-indexer",
   "scripts": {
-    "api": "node rest-api/server.js",
+    "api": "node api/server.js",
     "docker:build": "docker build -t ethereum-transaction-indexer .",
     "check": "dependency-check . && dependency-check . --unused --no-dev --ignore-module winston --ignore-module winston-papertrail --ignore-module winston-sentry-transport --ignore-module websocket-reconnector",
     "check:outdated": "npm-check -p -i husky -i eslint-config-plugin:*",
@@ -35,6 +35,7 @@
     "sinon": "7.2.4",
     "socket.io": "2.2.0",
     "web3": "1.0.0-beta.37",
+    "web3-utils": "1.0.0-beta.37",
     "websocket-reconnector": "1.1.1",
     "winston": "2.4.4",
     "winston-papertrail": "1.0.5",
